requirements.txt
opencv-python
numpy
pillow
google-genai
python-dotenv
ultralytics
transformers
torch
gradio
